+++ extensions/jetbrains/src/recentMenu.tsx
@@ -25,13 +25,6 @@ export default function ProjectList(): React.JSX.Element {
         />
       </MenuBarExtra>
     );
-  } else if (!toolboxApp.isV2) {
-    return (
-      <MenuBarExtra isLoading={isLoading} icon={menuIcon}>
-        <MenuBarExtra.Item title={"Wrong Jetbrains Toolbox version, please use V2"} />
-        <MenuBarExtra.Item title={`Current ToolBox version: ${toolboxApp.version}`} />
-      </MenuBarExtra>
-    );
   }
   return (
     <MenuBarExtra isLoading={isLoading} icon={menuIcon}>